requirements-backend.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
stripe>=9.0.0
apscheduler>=3.10.0
httpx>=0.27.0
python-multipart>=0.0.9
anthropic>=0.25.0
